Create example_slurm_script.sh
This commit adds the SLURM batch script `example_slurm_script.sh`, which serves as an example job submission script. The script is designed to run a job on the queue and sets various queuing parameters using #SBATCH directives. It includes commands to print out the nodes the job is running on and the number of cores per node. The script can be used as a template for submitting batch jobs to the SLURM job scheduler.